+++ pysui/sui/sui_grpc/suimsgs/sui/rpc/v2/__init__.py
@@ -223,7 +223,7 @@ if TYPE_CHECKING:
     from betterproto2.grpclib.grpclib_client import MetadataLike
     from grpclib.metadata import Deadline
 
-_COMPILER_VERSION = "0.9.0"
+_COMPILER_VERSION = "0.10.1"
 betterproto2.check_compiler_version(_COMPILER_VERSION)